Fixed word relative frequency in Austen novels
Relative word frequencies in Austen novels is now correctly computed by
using the denominator of total words before executing inner_join() with
Bronte or Wells novels. Now a word's relative frequency in the Austen
novels will be the same whether in the plot w/r/t Bronte or w/r/t Wells.

It also seems that the Austen words (in object tidy_books) have the same
underscores-for-italics UTF-8 encoding issues as in the Gutenberg texts,
so str_extract() is also applied to the Austin works. A consequence is
that both datasets in the inner_join() have some words that are simply NA
(as a result of chapter numbers that have been stripped to NA) and these
NA need to be removed before plotting.